etl/requirements.txt
psycopg2-binary>=2.9
requests==2.31.0
APScheduler==3.10.4
python-dotenv==1.0.1